styl: :art: split applicationRouting into components
